Rename _rust module to submodule-specific name
