+++ arro3-compute/Cargo.toml
@@ -13,7 +13,7 @@ rust-version = "1.75"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "_rust"
+name = "_compute"
 crate-type = ["cdylib"]
 
 [dependencies]
