+++ arro3-compute/pyproject.toml
@@ -14,7 +14,7 @@ classifiers = [
 
 [tool.maturin]
 features = ["pyo3/extension-module"]
-module-name = "arro3.compute._rust"
+module-name = "arro3.compute._compute"
 python-source = "python"
 
 [tool.poetry]
