+++ arro3-compute/python/arro3/compute/__init__.py
@@ -1,4 +1,4 @@
-from ._rust import *
-from ._rust import ___version
+from ._compute import *
+from ._compute import ___version
 
 __version__: str = ___version()
