+++ arro3-compute/src/lib.rs
@@ -14,7 +14,7 @@ fn ___version() -> &'static str {
 }
 
 #[pymodule]
-fn _rust(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
+fn _compute(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(___version))?;
 
     m.add_wrapped(wrap_pyfunction!(cast::cast))?;
