+++ arro3-core/Cargo.toml
@@ -13,7 +13,7 @@ rust-version = "1.75"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "_rust"
+name = "_core"
 crate-type = ["cdylib"]
 
 [dependencies]
