+++ arro3-core/python/arro3/core/_core.pyi
