+++ arro3-core/src/lib.rs
@@ -11,7 +11,7 @@ fn ___version() -> &'static str {
 
 /// A Python module implemented in Rust.
 #[pymodule]
-fn _rust(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
+fn _core(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(___version))?;
 
     m.add_class::<pyo3_arrow::PyArray>()?;
