+++ arro3-io/pyproject.toml
@@ -14,7 +14,7 @@ classifiers = [
 
 [tool.maturin]
 features = ["pyo3/extension-module"]
-module-name = "arro3.io._rust"
+module-name = "arro3.io._io"
 python-source = "python"
 
 [tool.poetry]
