+++ arro3-io/python/arro3/io/_io.pyi
